tools/doc-generator/requirements.txt
# FormalMath 文档自动生成系统依赖

# 核心依赖 (Python 3.8+ 内置)
# - pathlib
# - json
# - re
# - ast
# - argparse
# - datetime
# - dataclasses
# - collections
# - typing

# 可选依赖
pyyaml>=6.0          # YAML文件支持
markdown>=3.4        # Markdown处理
jinja2>=3.1          # 模板引擎（预留）

# PDF导出（可选）
# weasyprint>=59.0   # HTML转PDF
# pypandoc>=1.11     # 文档格式转换

# 开发依赖（可选）
# pytest>=7.0        # 测试框架
# black>=23.0        # 代码格式化
# mypy>=1.0          # 类型检查
